src/test/java/it/colletta/ApartmentTest.java
/* ==============================================
 * Copyright 2003 Mort Bay Consulting Pty Ltd. All rights reserved.
 * Distributed under the artistic license.
 * Created on 16/04/2004
 * $Id: ApartmentTest.java,v 1.4 2006/01/21 17:21:00 gregw Exp $
 * ============================================== */
 
package it.colletta;

import com.mortbay.iwiki.YyyyMmDd;

import junit.framework.TestCase;

/* ------------------------------------------------------------------------------- */
/** 
 * 
 * @version $Revision: 1.4 $
 * @author gregw
 */
public class ApartmentTest extends TestCase
{
    /* ------------------------------------------------------------------------------- */
    /** Constructor.
     * 
     */
    public ApartmentTest()
    {
    }
    
    public void testSeason()
    {

        // Low Season:  from the 10-01-04 till the 02-04-04, from the 16-10-04 till the 17-12-04
        // Mid Season:  from the 03-04-04 till the 09-04-04, from the 17-04-04 till the 11-06-04,
        //              from the 11-09-04 till the 15-10-04
        // High Season: from the 01-01-04 till the 09-01-04, from the 10-04-04 till the 16-04-04,
        //              from the 12-06-04 till the 10-09-04, from the 18-12-04 till the 07-01-05

        YyyyMmDd[] low = {
                new YyyyMmDd(2004,1,10),new YyyyMmDd(2004,4,2),
                new YyyyMmDd(2004,10,16),new YyyyMmDd(2004,12,17),
        };
        
        YyyyMmDd[] mid = {
                new YyyyMmDd(2004,4,3),new YyyyMmDd(2004,4,9),
                new YyyyMmDd(2004,4,17),new YyyyMmDd(2004,6,11),
                new YyyyMmDd(2004,9,11),new YyyyMmDd(2004,10,15),
        };

        YyyyMmDd[] high = {
                new YyyyMmDd(2004,1,1),new YyyyMmDd(2004,1,9),
                new YyyyMmDd(2004,4,10),new YyyyMmDd(2004,4,16),
                new YyyyMmDd(2004,6,12),new YyyyMmDd(2004,9,10),
                new YyyyMmDd(2004,12,18),new YyyyMmDd(2005,1,5), 
        };
        
        for (int i=0;i<low.length;i+=2)
        {
            YyyyMmDd ymd=new YyyyMmDd(low[i]);   
            while (ymd.before(low[i+1]) || ymd.equals(low[i+1]))
            {
                assertEquals(ymd.toString(),Apartment.LOW,Apartment.getSeason(ymd));
                ymd.addDays(1);
            }
        }
        
        /* TODO failing - no idea why
        for (int i=0;i<mid.length;i+=2)
        {
            YyyyMmDd ymd=new YyyyMmDd(mid[i]);   
            while (ymd.before(mid[i+1]) || ymd.equals(mid[i+1]))
            {
                assertEquals(ymd.toString(),Apartment.MID,Apartment.getSeason(ymd));
                ymd.addDays(1);
            }
        }
        
        for (int i=0;i<high.length;i+=2)
        {
            YyyyMmDd ymd=new YyyyMmDd(high[i]);   
            while (ymd.before(high[i+1]) || ymd.equals(high[i+1]))
            {
                assertEquals(ymd.toString(),Apartment.HIGH,Apartment.getSeason(ymd));
                ymd.addDays(1);
            }
        }
        */
        
    }
    
}
